\name{lrgs-package}
\alias{lrgs-package}
\alias{lrgs}
\docType{package}
\title{
Linear Regression by Gibbs Sampling
}
\description{
Implements a Gibbs sampler to do linear regression with multiple covariates, multiple responses, Gaussian measurement errors on covariates and responses, Gaussian intrinsic scatter, and a covariate prior distribution which is given by either a Gaussian mixture of specified size or a Dirichlet process with a Gaussian base distribution.
}
\details{
\tabular{ll}{
Package: \tab lrgs\cr
Type: \tab Package\cr
Version: \tab 0.4.2\cr
Date: \tab 2015-09-04\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}

See help for Gibbs.regression
}
\author{
Adam Mantz <amantz@slac.stanford.edu>
}

\keyword{ package }
